fix: workdoc list not support align (#8703)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added support for the `Workdoc` field type in the list of
field types in the `TypeSelect` component, enhancing the functionality
of the multi-grid feature.
- New Feature: Improved the user interface of the editor by dynamically
showing/hiding alignment options based on the active editor state. The
`AlignSelect` component will now be hidden when the editor is active in
task list, bullet list, ordered list, or blockquote modes.
- Style: Modified the class name of a component in the image resizer
function from 'ProseMirror-selectednode' to 'image-selected',
potentially affecting the styling and behavior of the component.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->