fix: link default prefix https (#8714)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->


<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Improved default prefix for links in the bubble menu by
automatically adding "https://" if the input link doesn't start with it.
- Refactor: Removed the `HorizontalRule` extension from the list of
default extensions and added a configuration option to disable document
history and drop cursor.
- Bug Fix: Updated the URL property of the `HocuspocusProvider` object
based on the environment, ensuring correct behavior in different
environments.
- Style: Modified CSS styles for buttons and text in the UI, enhancing
the appearance of the work document buttons.
- New Feature: Introduced a new field type, `FieldType.Workdoc`, in the
`ChangesetItemActionBase` component for a specific condition.
- Style: Added a new CSS class with a minimum height for form
containers, improving the visual layout.
- New Translation: Added a new translation key for a tooltip related to
editing a form field for a work document.
- Refactor: Added conditional logic for the mobile view of the form
editor, returning a specific component based on the `isMobile` variable.
- Style: Added a default empty string value for the `title` property in
the length calculation, preventing errors when `item.text` and `(item as
any).title` are undefined or null.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->