fix: workdoc image style (#8736)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Style: Modified the CSS style of the `.workdocBtn` class, changing the
`display` property and adding a `max-width` property to affect the
visual layout and behavior of a button.
- Style: Added a new CSS rule to the `.left` class, setting the
`margin-right` property to affect the styling of an image in the
workdoc.
- Style: Made various changes to the CSS styles of the workdoc image,
including modifying padding, margin, font-weight, background color, and
adjusting the width, height, background color, border, and position of
resize triggers.
- Refactor: Modified the `Resizer` component to always render an `<img>`
element instead of conditionally rendering a `<figure>` element.
- Bug Fix: Modified the condition inside an `if` statement in the
`FloatingMenu` component to allow the condition to pass when
`textContent` is `'、'` as well.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->