fix: form create new workdoc (#8769)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->

<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Refactor: Removed the `showWorkdoc` state and its usages from multiple
components, as well as updated the import statement for
`useAppSelector`. The code now directly calls the `onSave` function with
a new document ID when `cellValue` is null, allowing the creation of a
new workdoc without showing the workdoc UI.
- Refactor: Changed the rendering order of elements in the canvas by
moving a `path` function call below a `label` function call in the
`CellHelper` class.
- Style: Added the color property to the CSS class ".drop-cursor" in the
`styles.module.less` file, setting the color of the cursor to the value
of the "--textCommonPrimary" variable.
- Chore: Removed import statements for `DragDrop` from two files and
commented out the lines where `DragDrop` was used in the
`default-extensions.tsx` file.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->