Refactor/training (#8802)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->

<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added a new feature for managing data sources and
training AI models. This includes the introduction of the `DataSource`
component, which renders configuration items and tables for data
sources. The `DataSourceModal` component provides a modal dialog for
selecting data sources with different tabs for options. The `Training`
component has been refactored and enhanced for training AI models.
- Style: Made CSS styling modifications to various components, including
`ButtonArea`, `SourceItem`, `StatusItem`, `Table`, `TimeItem`,
`DataSourceModal`, `ConfigItem`, and `Training`. These changes are
purely cosmetic and do not affect the logic or functionality of the
code.
- Refactor: Introduced a new component called `ConfigItem`, which is a
configurable item with a title, description, and optional child
components. It supports a switch control for configuration toggling and
provides a link to additional documentation.
- Bug Fix: Fixed CSS styles in the `Training` component by removing
unnecessary padding and modifying background colors using custom CSS
variables.
- Documentation: Added new translation keys related to training data
sources for displaying text in the user interface.
- Refactor: Added a new function `convertStatus2Code` in the
`Training/utils.ts` file to map status strings to corresponding codes
from the `TrainingStatus` enum.
- New Feature: Introduced a new component called `HistoryDetailTable`
that displays a table with historical training data. It utilizes the
`ConfigItem` component to display configuration details.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->