fix: notify url (#8878)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Expanded options for providing a URL in toast
notifications by allowing the `url` property to accept a string or an
object with a `text` property.
- Refactor: Modified the `stringToActions` function to support a new
type of `url` parameter, allowing it to accept an object with a `text`
property. The function now extracts the URL string accordingly before
invoking the appropriate action based on the `actionStr`.
- Refactor: Made changes to import statements, component rendering, and
introduced a new `VikabyWithTheme` component wrapped in a
`ThemeProvider`. The `showVikaby` function now includes additional
conditions before rendering the `Vikaby` component. These changes do not
significantly impact logic or functionality.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->