fix: added join human machine verification (#8885)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: The `joinViaSpace` function in the `api.org.ts` file now
accepts an additional parameter `data`, which is included in the payload
of the POST request made by the function.
- Bug Fix: The `LinkConfirm` component in `link_confirm.tsx` now
includes a check for `isLogin` before calling the `join` function. If
the user is logged in, the `execNoTraceVerification` function is
executed before calling `join`.

<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->