fix: workdoc image loading (#8908)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Style: Added a new CSS keyframe animation called
`imageLoadingRotation` to create a loading effect for images.
- Refactor: Modified the `NodeSelect` component by importing the `last`
function from the `lodash` library, updating the logic for determining
the active item and active list, adding a new variable
`listActiveFnName`, modifying the `onSelected` event handler to handle
list active state, and refactoring the options rendering.
- Refactor: Added a new function `list2Text` to convert the active list
type in the editor to its corresponding command name.
- Style: Updated CSS styles of an image loading feature, including
increasing the border width of an icon, adding padding to a code block,
and adding a new loading animation to the image loading icon.
- New Feature: Added a loading icon to the image placeholder in the
`UploadImagePlaceholderPlugin` function.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->