fix: workdoc color ui (#8925)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Style: Added `white-space: inherit !important;` CSS property to the
`color` class within a specific div in the
`activity_item/style.module.less` file.
- Style: Modified the `ColorSelect` component in the
`color-select/index.tsx` file by adding a border style to the list item
when it is the active color item. The border color is set to the
corresponding color value.
- Style: Removed the `border-color` property from the `.colorActive` CSS
class in the `color-select/styles.module.less` file.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->