Feat/third im seat limit (#9031)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
third im seat limit 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->

<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added functionality to retrieve the user ID based on the
provided OpenID in the organization module.
- New Feature: Modified the `checkSeatOverLimitAndSendNotify` function
in the space module to handle multiple user IDs at once.
- Bug Fix: Updated the `handleFeishuEvent` method in the GM service
module to include an additional argument in the
`handleTenantContactData` method call.
- New Feature: Added a static method to create a message card for
notifying that the seat limit has been reached in the DingTalk
integration module.
- New Feature: Added a static method to create a card message for
notifying that the seat limit has been reached in the Feishu integration
module.
- New Feature: Added a function to generate a message for when the seat
limit is exceeded in the WeCom integration module.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->