fix: image resize limit max width (#9060)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: The `WorkDoc` field type is now included in the
`fieldSequence` array in the `TypeSelect` component, allowing users to
select and work with this field type.
- Style: The CSS code has been updated to disable interaction with child
elements of the `.link` class, preventing any unintended actions on
those elements.
- Bug Fix: The maximum width of resized images in the `ImageResizer`
component is now constrained to the width of the `.ProseMirror` element,
ensuring that images do not exceed the available space.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->