Hotfix/cola (#9083)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Fixed an issue in the `expand_upgrade_space.tsx` file where
the `onModalClose` function was not properly clearing the query
parameters. Now, the function uses the `Router.push` method to navigate
to the workbench page with cleared query parameters.
- Refactor: In the `workspace.tsx` file, refactored the
`resumeUserHistory` function to check the `appType` property of the
`social` object and return early if it is either 1 or 2. This improves
the logic and ensures that the function behaves correctly based on the
app type.
- New Feature: Added the ability to retrieve and access the current
space information in the `workspace.tsx` file. The `useAppSelector`
function is used to fetch the space information, and the `spaceInfo`
variable is introduced to store it. This allows for further
customization and utilization of the space information in the code.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->