fix: konva workdoc style (#9075)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->

<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added a new constant `GRID_CEL_ICON_GAP_SIZE` with a
value of 8 to represent the gap between icons in a grid cell.
- Bug Fix: Adjusted the calculation of `maxTextWidth` and updated the
padding value for the background shape in the `CellHelper` class,
potentially improving the rendering of cell values and the appearance of
the background shape.
- Refactor: Removed the unused `setAutomationState` function from the
destructuring assignment in the `useAtom` hook usage in
`RenderModalBase` and `DescriptionModal` components.
- Documentation: Updated the interface `IQuery` by adding two
properties, `recordId` and `fieldId`, which may affect the external
interface of the code.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->