fix: create workdoc link (#9104)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Modified the deletion behavior in the floating menu to delete
both the forward slash and the selected text, improving the range of
characters to be deleted.
- Refactor: Removed the `DocumentWithTitle` extension in the editor core
and directly extended the `Document` class with a content specification,
without altering the external interface or behavior.
- Style: Added the `display: inline-block;` property to the
`.img-placeholder` CSS class in the editor core styles, allowing it to
be displayed as an inline block.
- New Feature: Updated the workdoc editor to include navigation and
selectors from `@apitable/core`, and replaced the `onStartEdit` function
to include a call to `Router.replace` with updated parameters.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->