Fix/fix automation select panel (#9235)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Refactor: The `getColorValue` function in `color.ts` is now exported
as a public function.
- Style: Modifications to the rendering of color groups in the UI
component `color_group.tsx`. Adjustments made to text color and
background color based on the selected color index. Removal of
commented-out code related to button styles.
- Refactor: Changes to the `ButtonFieldItem` and `ButtonItem`
components. Introduction of a new styled component `StyledBgBox` to
handle background color changes on hover and click. Update of the
`StyledBox` component to use `StyledBgBox` and accept a `color` prop.
Removal of the `colorList` variable in `ButtonItem`. Update of the
`StyledBox` component in `ButtonItem` to use `StyledBgBox` and accept a
`backgroundColor` prop.
- Bug Fix: Modification of the `useButtonFieldValid` function. Update of
the logic for finding the `buttonFieldTriggerId` by including an
additional condition to check for the `button_clicked` endpoint.
- Style: Removal of the right margin from the `.errorMsg` class in the
CSS file `style.module.less`.
- New Feature: Introduction of a new value `'button_clicked'` to the
`EnumTriggerEndpoint` type.
- New Feature: Addition of a new case `'button_clicked'` to a switch
statement in the `RobotTriggerBase` component. Modification of the
`triggerTypeOptionsWithoutButtonIsClicked` variable to exclude both
`'button_field'` and `'button_clicked'` when the automation scenario is
`'datasheet'`.
- Refactor: Removal of imports from external modules, removal of hover
and active styles, and modification of the `onClick` event handler in
the `StyledBox` component in `robot_readonly_card.tsx`.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->