fix: new field tag (#9299)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added two new optional fields, `isBeta` and `isNew`, to
the `IFieldTypeCollection` interface in `field_types.ts`. These fields
allow for tagging specific field types as beta or new.
- Style: Introduced two new CSS classes, `.beta` and `.new`, in
`styles.module.less` for displaying beta and new tags with custom
styling in the type select component.
- New Feature: Updated the `FieldTypeDescriptionMap` object in
`type_select.tsx` to include the `isBeta` and `isNew` fields. This
enables the conditional rendering of beta and new tags in the type
select component.
- Refactor: Modified the `tippyOptions` prop in
`floating-menu/index.tsx` to enhance the behavior and appearance of the
floating menu component.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->