Feature/fix fix button (#9361)
@coderabbitai:ignore

close #9362 

<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Introduce `AutomationTaskStatus` type to define possible
statuses of automation tasks. Update `automationTaskMap` atom to use
`AutomationTaskStatus` type instead of boolean values.
- New Feature: Add `taskStatus` prop to `ButtonItem` component to handle
automation task status. Replace usages of `isLoading` prop with
`taskStatus` in `ButtonItem` to conditionally render UI elements based
on task status. Use `useRef` hook to store a reference to the timeout
ID. Set a timeout to reset the task status after 1 second when the
automation task is successful.
- New Feature: Modify `runAutomationButton` function to provide feedback
on the success of the button action. Add a new parameter to the
`callback` function to indicate the success of the button action. Call
`callback` function with the value of `respData.success` if it exists,
otherwise pass `false`.
- Refactor: Remove console.log statement that printed the value of the
`field` variable.
- Bug Fix: Add null check for the `data` parameter in the `check`
function in `use_button_field_valid.ts`. If `data` is null, return an
object with `fieldId` set to `button.id` and `result` set to `false`.
Ensure the function handles the case when `data` is not provided.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->