fix: view chart line (#9368)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Modified the `calcCellValueAndString` function in the
`cell_calc.ts` file to address an issue related to chart line view. The
`ignoreCache` flag is now set to `true` when the field type is
`FieldType.CreatedBy`.
- Refactor: Adjusted the height calculations for fields in a card view
in the `useCardHeight` function of the `use_card_height.tsx` file.
Changes include modifying the values of `FIRST_FIELD_HEIGHT`,
`FIELD_TITLE_HEIGHT`, and `FIELD_PADDING`, as well as adjusting the
calculation for total card height to include padding.
- New Feature: Added entries for the `WorkDoc` and `OneWayLink` field
types in the `FIELD_HEIGHT_MAP`, `FIELD_HEIGHT_VIRTUAL_MAP`, and
`FIELD_HEIGHT_MAP_MOBILE` objects in the `constant.ts` file. This allows
for proper rendering of these field types in the UI.
- Chore: Removed duplicate import statement for the `useAppSelector`
function in the `card_body.tsx` file to avoid duplication.
- Style: Added a new global style for the class `.expandWorkdoc` in the
`style.module.less` file, affecting the appearance of the chart line
view.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->