Revert "fix: lookup ignore cache (#9537)" (#9589)
This reverts commit d7fa171e2aee07a61629339ae7367fb380876a57.


# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- Bug Fix: Modified the `calcCellValueAndString` function in
`cell_calc.ts` to change the condition for ignoring cache fields.
Instead of using the `IGNORE_CACHE_FIELDS` constant, it now checks if
the field type is `FieldType.CreatedBy`. This may affect the behavior of
the code related to caching.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->