fix: workdoc hide copy field (#9626)
# Human says

As a human, what have you done for this pull request?
Why? What? How?
<!-- 如果这个Pull Request有对应的Issue，你可以复制Issue的相关信息。 -->
<!-- 
# Why?
> 对应哪个issue?
> 为什么要这个pull request？
> 背景故事或原因是怎样的？
> 你理解的业务需求是怎样的？ 
-->

<!-- 
# What?
> 这是一个什么Pull Request？
> 描述一下是什么？
> 对哪些人有好处？ 
-->

<!-- 
# How?
> 大概描述下，如何具体实现的这个Pull Request？ 
-->


# AI says

<!-- AI auto review added here -->


<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added a condition to hide a field menu option in the
datasheet component based on the type of the field.
- New Feature: Implemented the useLocalStorageState hook from the ahooks
library to manage the state of the table of contents in the editor. The
table of contents can now be opened or closed, with the default value
set to open.
- Bug Fix: Modified the logic for updating the table of contents in the
editor to exclude headings with empty text content.
- Refactor: Conditionally render the WorkdocCollaborators component in
the workdoc editor based on the device type. The component is now hidden
on mobile devices.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->